feature(macro): added macro for enum<->u8

+++ src/traits/taker.rs
@@ -7,7 +7,7 @@ use std::convert::{TryFrom, TryInto};
 pub trait Taker: Send + Sync
 {
     fn take_hinted<T: FromValue>(&mut self, name: &str) -> Result<T, FromRowError>;
-    fn take_enum<T: TryFrom<usize>>(&mut self, name: &str) -> Result<T, FromRowError>;
+    fn take_enum<T: TryFrom<u8>>(&mut self, name: &str) -> Result<T, FromRowError>;
     fn take_date_time(&mut self, name: &str) -> Result<DateTime<Utc>, FromRowError>;
     fn take_date_time_option(&mut self, name: &str) -> Result<Option<DateTime<Utc>>, FromRowError>;
 }
@@ -18,9 +18,9 @@ impl Taker for Row
         self.take::<T, &str>(name).ok_or_else(|| FromRowError(self.clone()))
     }
 
-    fn take_enum<T: TryFrom<usize>>(&mut self, name: &str) -> Result<T, FromRowError>
+    fn take_enum<T: TryFrom<u8>>(&mut self, name: &str) -> Result<T, FromRowError>
     {
-        self.take::<usize, &str>(name)
+        self.take::<u8, &str>(name)
             .ok_or_else(|| FromRowError(self.clone()))?
             .try_into()
             .map_err(|_| FromRowError(self.clone()))
